Split components
Now checks for instance before modifying text, stopping null reference
exceptions

InformationText -> InfoPanel and InfoOverlay
GameObject name/editor fields updated
Documentation updated

Fields and methods updated
Callers updated
