+++ Holojam/Prefabs/Holojam.prefab
@@ -15,7 +15,8 @@ GameObject:
   - component: {fileID: 114653569414185928}
   - component: {fileID: 114000013142514114}
   - component: {fileID: 114694228294611138}
-  - component: {fileID: 114256559976128022}
+  - component: {fileID: 114455135630302756}
+  - component: {fileID: 114735248688510034}
   m_Layer: 0
   m_Name: Holojam
   m_TagString: Untagged
@@ -70,7 +71,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ce1d2ec38187349efbbc678df0de9a00, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &114256559976128022
+--- !u!114 &114455135630302756
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
@@ -78,9 +79,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 160812}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6776f27b5c65a4811ada7286a064ece9, type: 3}
+  m_Script: {fileID: 11500000, guid: 6fb0eb229cd6e4f11b5e1201c2c08e89, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  display: 0
+  toggleKey: 96
 --- !u!114 &114615662255141442
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -117,6 +120,17 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 1834260139, guid: 1362ee83563f0458bbbc776de423bfe8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &114735248688510034
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 160812}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 78ee8f34053d447f79b4fdeb4b1890be, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &114736255769847042
