+++ Holojam/Tools/InfoPanel.cs.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 6776f27b5c65a4811ada7286a064ece9
-timeCreated: 1490925998
-licenseType: Free
+guid: 6fb0eb229cd6e4f11b5e1201c2c08e89
+timeCreated: 1492291036
+licenseType: Pro
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
