+++ Holojam/Tools/Overlay.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 78ee8f34053d447f79b4fdeb4b1890be
+timeCreated: 1492291036
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
